+++ apps/app_web/assets/css/anacounts.css
@@ -11,5 +11,4 @@
 @import "./components/heading.css";
 @import "./components/icon.css";
 @import "./components/list.css";
-@import "./components/page-header.css";
 @import "./components/toggle-nav.css";
