+++ apps/app_web/lib/app_web.ex
@@ -91,6 +91,7 @@ defmodule AppWeb do
       # Import LiveView and .heex helpers (live_render, live_patch, <.form>, etc)
       import Phoenix.LiveView.Helpers
       import AppWeb.LiveHelpers
+      import AppWeb.ComponentHelpers
 
       # Import basic rendering functionality (render, render_layout, etc)
       import Phoenix.View
